requirements.txt
amqp==5.2.0
APScheduler==3.10.4
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
celery-redbeat==2.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.0
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.1.1
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
responses==0.24.1
schedule==1.2.1
six==1.16.0
SQLAlchemy==2.0.23
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
xecd-rates-client==1.0.0
